+++ src/Builders/RequestBuilder/GpEcom/GpEcomManagementRequestBuilder.php
@@ -65,6 +65,13 @@ class GpEcomManagementRequestBuilder extends GpEcomRequestBuilder implements IRe
         }
 
         $request->appendChild($xml->createElement("orderid", $orderId));
+        if (isset($builder->surchargeAmount)) {
+            $surchargeAmount = $xml->createElement("surchargeamount", preg_replace('/[^0-9]/', '', sprintf('%01.2f', $builder->surchargeAmount)));
+            if (!empty($builder->creditDebitIndicator)) {
+                $surchargeAmount->setAttribute("type", strtolower($builder->creditDebitIndicator));
+            }
+            $request->appendChild($surchargeAmount);
+        }
         $request->appendChild($xml->createElement("pasref", $builder->transactionId ?? ''));
 
         // rebate hash