+++ src/Entities/Enums/UserType.php
@@ -7,4 +7,5 @@ use GlobalPayments\Api\Entities\Enum;
 class UserType extends Enum
 {
     const MERCHANT = 'MERCHANT';
+    const PAYER = 'PAYER';
 }