+++ test/Integration/Gateways/GpApiConnector/GpApiBNPLTest.php
@@ -879,7 +879,7 @@ class GpApiBNPLTest extends TestCase
     private function setCustomerData(): Customer
     {
         $customer = new Customer();
-        $customer->id = "12345678";
+        $customer->key = "12345678";
         $customer->firstName = 'James';
         $customer->lastName = 'Mason';
         $customer->email = 'james.mason@example.com';