+++ test/Integration/Gateways/GpApiConnector/PartnershipModeTest.php
@@ -465,7 +465,7 @@ class PartnershipModeTest extends TestCase
         $address->countryCode = "US";
 
         $customer = new Customer();
-        $customer->id = "e193c21a-ce64-4820-b5b6-8f46715de931";
+        $customer->key = "e193c21a-ce64-4820-b5b6-8f46715de931";
         $customer->firstName = "James";
         $customer->lastName = "Mason";
         $customer->dateOfBirth = "1980-01-01";